unique test for all fluid entities
